fix: fan icon giant-size animation (frame+transform actually fixed); no-icon default = frosted glass like ball; app list shows real icons (1.2.10)

+++ Preferences/FUSettingsController.m
@@ -332,7 +332,7 @@ static const NSInteger kFUMaxEntries   = 6;
             if ([bid isEqualToString:@"com.apple.Preferences"]) continue;
             UIImage *icon = nil;
             if (uiImg && [uiImg respondsToSelector:iconSel]) {
-                NSInteger fmt = 2; CGFloat scale = (UIScreen.mainScreen ? UIScreen.mainScreen.scale : 2.0f);
+                int fmt = 2; CGFloat scale = (UIScreen.mainScreen ? UIScreen.mainScreen.scale : 2.0f);
                 NSInvocation *inv = [NSInvocation invocationWithMethodSignature:
                     [uiImg methodSignatureForSelector:iconSel]];
                 [inv setTarget:uiImg]; [inv setSelector:iconSel];
